requirements.txt
pymupdf==1.23.7
camelot-py==0.11.0
pandas==2.2.2
pdfplumber==0.11.0
tabulate==0.9.0
numpy==1.26.4
paddleocr==2.7.0.3
paddlepaddle==2.6.1
opencv-python==4.9.0.80
Pillow==10.2.0